requirements.txt
line-bot-sdk
flask
gunicorn
Pillow==8.2.0